Mobile usability
Fixed number input on iOS to trigger the number input button appropriately, as excellently explained on the following:
http://danielfriesen.name/blog/2013/09/19/input-type-number-and-ios-numeric-keypad/

This has the unfortunate side effect of bricking the "number" input validation on some desktop browsers, but not Chrome.  With that said given this declaration it is not possible to submit bad input using a modern browser right now.  Users are still guided to do the right thing.