Return named tuple from search_port
This allows for the result to be used in an object-like way, rather than "guessing" the data based on position in an ordered list.  Technically it's the same order from the same list, but this generates the named tuple in a systematic way, allowing for mnemonic reuse.

in web.py, this now passes the entire "search_result" object to the template page.