Disallow submission until input field is filled
Prevents the page from being submitted if input is null.  This makes use of the native "required" HTML5 attribute.

Using "required" implied changing the <input> to a true <button>, but this is more HTML5 freindly too.

An unexpected change was that the adoption of <button> made the button appear IOS styled on IOS devices, consider the use of the style below to address it in the future if consistent style is important.  Right now it is ok to default on mobile.

  -webkit-appearance: none;
  -moz-appearance: none;