Even better data quality, incredible speed
Although a point of potential contentious debate, not committing each transaction in setup saves MINUTES of processing time during build activity.  Other changes were made to ensure consistency, and the positive speed benefit was not resultant in lesser data quality.  All records appear in the SQLite DB without NULL entry.

If the database was highly transactional and updating records as a result of user activity, I might be more concerned.  But because all of this happens prior to user exposure and is purely backend, at 16K records and no concurrent connections this is probably relatively safe.