Address permissions issue due to sqlalchemy implementation
Although elegant, the sqlalchemy improvement broke the docker build due to bad permissions.  the flask-user did not have enough permissions to create the sqlite database.  This wasn't caught in testing because I believe the sqlite database was copied into the docker image (which has also been addressed).

After trying a few permission workarounds, the one i am most comfortable with is a change to a multi-stage docker build.  This also required some small modifications to the sqlalchemy implementation.