Fixed priveleges, some more
Permissions were still broken, now this time fixed for good.

As it turns out, the flask-user COULD create the database, but not in the /usr/local/src/knowyourports/ directory.

The learning here is that sqlite:/// (three slashes; relative) is relative to the script, not the current working directory as I thought.

I could have done away with the multi-stage build as a result of this, but it's slick enough that I want to keep it for now, until there's a good enough reason not to.