removing container scan for now
the Azure container scan was an experiment and a good one, but it is slightly noisy given debian vulnerability posture as well as the fact that this is a pre-release action.
It is not clear when the V1 of this action will be published.

A number of best practices were adoped as a result of this.  Container scanning will return to this project.